API Documentation

Type hierarchy

CitableAnnotations.AbstractAnnotationSet — Type

Abstract type for a citable annotation.

CitableAnnotations.CitableCommentary — Type

Annotations by one text on another text.

Metadata about annotation sets

CitableAnnotations.annotatedtype — Function

Find Julia type of annotated objects.

annotatedtype(anns)

In a CitableCommentary, texts commented on are identifed by CtsUrn.

annotatedtype(comm)

In TextOnPage annotations, annotated pages are identifed byCite2Urn`.

annotatedtype(idx)

CitableAnnotations.annotatingtype — Function

Find Julia type of annotating source.

In a CitableCommentary, commenting texts are identifed by CtsUrn.

annotatingtype(comm)

In TextOnPage annotations, texts are identifed by CtsUrn.

annotatingtype(idx)

CitableAnnotations.annotated — Function

Find unique list of annotated objects.

Find unique set of texts commented on.

annotated(comm)

URNs identify versions or exemplars of texts without passage references.

Find unique set of pages.

annotated(idx)

CitableAnnotations.annotators — Function

Find unique list of annotating sources.

Find unique set of texts commenting on other texts.

annotators(comm)

URNs identify versions or exemplars of texts without passage references.

Find unique set of texts appearing on pages.

annotators(idx)

URNs identify versions or exemplars of texts without passage references.